pyproject.toml
[project]
name = "judiciary-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django==4.2",
    "djangorestframework==3.14.0",
    "django-filter==23.5",
    "django-cors-headers==4.3.1",
    "django-simpleui>=2023.0",
    "python-json-logger>=2.0.7",
    "django-ckeditor>=6.7.0",
    "requests>=2.31.0",
#    "redis>=5.0.0",
#    "django-redis>=5.4.0",
    "cos-python-sdk-v5>=1.9.30",
    "Pillow>=10.0.0",
    "django-import-export>=3.3.0",
    "openpyxl>=3.1.0",
    "gunicorn>=21.0.0",
    "whitenoise>=6.6.0",
]
